product admin

+++ src/routers/productRouter.ts
@@ -170,7 +170,7 @@ routerProduct.put('/update/:id', authenticateToken, async (req: Request, res: Re
 // delete product
 
 routerProduct.delete('/:id', authenticateToken, async (req: Request, res: Response)=>{
-
+  
 
   try {
 
@@ -191,25 +191,43 @@ routerProduct.delete('/:id', authenticateToken, async (req: Request, res: Respon
         'message': 'Prodotto non trovato'
       });
     }
-    
+    const authHeader = req.headers["authorization"];
+    const token = authHeader && authHeader.split(" ")[1];
+    const userFound = await db.get('SELECT userID FROM auth WHERE token = ?',[token]);
 
-    const result = await db.run(
-      'DELETE from products WHERE id = ?',[id]
-    );
-
-    if (result.changes && result.changes > 0) {
-      return res.status(200).json({
-        'message': 'Prodotto eliminato con successo',
-        'productId': id
-      });
-    } else {
-      return res.status(400).json({
-        'message': 'prodotto non eliminato'
+    if(!userFound){
+      return res.status(404).json({
+        'message': 'utente non trovato'
       });
     }
 
+    const userId = userFound.userId;
 
+    const isAdmin = userFound.is_admin;
 
+    if(isAdmin==1)
+    {
+      const result = await db.run(
+        'DELETE from products WHERE id = ?',[id]
+      );
+  
+      if (result.changes && result.changes > 0) {
+        return res.status(200).json({
+          'message': 'Prodotto eliminato con successo',
+          'productId': id
+        });
+      } else {
+        return res.status(400).json({
+          'message': 'prodotto non eliminato'
+        });
+      }
+    }
+    else
+    {
+      return res.status(401).json({
+        'message': 'Non hai i permessi per l\'operazione'
+      });
+    }
   } catch (err) {
     if (err instanceof Error) {
       return res.status(500).json({
